created new directory for monday c project and added readme

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -8,9 +8,9 @@ int main(void)
 {
 	int counter = 2;
 
-	double a = 1;
-	double b = a + 1;
-	double c = a + b;
+	float a = 1;
+	float b = a + 1;
+	float c = a + b;
 
 	printf("%.0f, ", a);
 	printf("%.0f, ", b);
